refactor utils paths
