+++ blocks/schedule/schedule.js
@@ -1,7 +1,7 @@
-import { append } from '../../scripts/dom-utils.js';
+import { append } from '../../scripts/utils/dom.js';
 import { readBlockConfig } from '../../scripts/lib-franklin.js';
 import { getScheduleData } from '../../scripts/services/ScheduleData.js';
-import { getSiteRoot } from '../../scripts/site-utils.js';
+import { getSiteRoot } from '../../scripts/utils/site.js';
 
 const dayIdPattern = /^#day-(\d)$/;
