+++ blocks/talk-detail-before-outline/talk-detail-before-outline.js
@@ -1,4 +1,4 @@
-import { append } from '../../scripts/dom-utils.js';
+import { append } from '../../scripts/utils/dom.js';
 
 /**
  * Talk Detail footer with back to schedule link.
